fix: get artifact and metafile paths from sls function object
Gets the paths to the metadata file and the zipped artifact from the
FunctionDefinitionHandler object exposed by the Serverless API. This
makes the plugin work correctly when e.g. the handler path is something
like `src/myFunctionHandler.main`.

Previously this was failing for me where e.g. the function name in
Serverless was 'myFunction' but the file path was
`myFunctionHandler.ts`.